KEYCLOAK-16755 ClearExpiredUserSessions optimization. Rely on infinispan expiration rather than Keycloak own background task.

+++ common/src/main/java/org/keycloak/common/util/Time.java
@@ -65,7 +65,7 @@ public class Time {
      * @param time Time in seconds since the epoch
      * @return Time in milliseconds
      */
-    public static long toMillis(int time) {
+    public static long toMillis(long time) {
         return time * 1000L;
     }
